Fix pagination and improve speed
Rather than getting all the PRs for the repo, and then selecting the ones that are earlier than the last commit_date, check the PRs each time I get a list. This way we do not need to get thousands of PRs, if we only need the first 30 or so.